+++ WebStore/Controllers/HomeController.cs
@@ -5,13 +5,7 @@ namespace WebStore.Controllers
 {
     public class HomeController : Controller
     {
-        private static readonly List<Employee> _Employees = new List<Employee>()
-        {
-            new Employee() { Id = 1, LastName = "Иванов", FirstName = "Иван", Patronomic = "иванович", Age = 27 },
-            new Employee() { Id = 2, LastName = "Петров", FirstName = "Петр", Patronomic = "Петрович", Age = 30 },
-            new Employee() { Id = 3, LastName = "Сидоров", FirstName = "Сидр", Patronomic = "Сидорович", Age = 27 }
-        };
-
+        
         public IActionResult Index()
         {
             //return Content("Data from new controller");
@@ -24,19 +18,8 @@ namespace WebStore.Controllers
         }
 
 
-        public IActionResult Employees()
-        {
-            return View(_Employees);
-        }
-
-        public IActionResult EmployeesInfo(int id)
-        {
-            var employee = _Employees.FirstOrDefault(o => o.Id == id);
-            
-            if (employee == null)
-                return NotFound();
-            return View(employee);
-        }
+       
+       
     }
