+++ WebStore/Views/Employees/Index.cshtml
@@ -1,17 +1,10 @@
 ﻿@model IEnumerable<WebStore.Models.Employee>
 
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Employees</title>
-</head>
-<body>
+<div class="container">
 
     <h3>Сотрудники</h3>
 
-    <table>
+    <table class="table">
         <thead>
             <tr>
                 <th>Id</th>
@@ -39,8 +32,5 @@
         </tbody>
     </table>
     <hr />
-	<a href="~/Home/Index">Главная страница</a>
-
-
-</body>
-</html>
+	<a  asp-controller="Home" asp-action="Index">Главная страница</a>
+</div>
