+++ WebStore/Views/Home/Index.cshtml
@@ -14,12 +14,6 @@
 				<partial name="Partial/_Category-tab"/><!--/category-tab-->
 					
 				<partial name="Partial/_Recommended_items"/><!--/recommended_items-->
-
-				
-					
-				
-					
-				
 					
 			</div>
 		</div>
